requirements.txt
asgiref==3.11.1
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==4.2.9
gunicorn==25.1.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
MarkupSafe==3.0.3
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
Pillow==10.1.0
pycparser==3.0
PyMySQL==1.1.2
pypdf==6.7.5
PyPDF2==3.0.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2026.1.post1
reportlab==4.4.10
requests==2.32.5
selenium==4.40.0
sendgrid==6.12.5
setuptools==69.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.6
wheel==0.43.0
whitenoise==6.5.0
wsproto==1.3.2
zipp==3.23.0
